Match aux_loss dtype to ce_loss in SemanticSegmentationLoss
`torch.tensor(0.0, ...)` defaulted to float32 regardless of the dtype
ce_loss/dice_loss end up in. Under AMP, this produced a loss_items
tensor whose dtype was implicitly promoted on stack, making it
brittle to future dtype tightening. Threading `dtype=ce_loss.dtype`
keeps the three loss components consistent.

+++ ultralytics/utils/loss.py
@@ -1373,8 +1373,8 @@ class SemanticSegmentationLoss(nn.Module):
         dice_loss = self._dice_loss(preds, masks)
         total = ce_loss + dice_loss
 
-        # Auxiliary cross-entropy loss.
-        aux_loss = torch.tensor(0.0, device=preds.device)
+        # Auxiliary cross-entropy loss. Match ce_loss dtype so torch.stack below succeeds under AMP.
+        aux_loss = torch.tensor(0.0, device=preds.device, dtype=ce_loss.dtype)
         if aux_logits is not None:
             if aux_logits.shape[2:] != masks.shape[1:]:
                 aux_logits = F.interpolate(aux_logits, size=masks.shape[1:], mode="bilinear", align_corners=False)
